--- a/src/arm64/bitmanipulation.h
+++ b/src/arm64/bitmanipulation.h
@@ -3,6 +3,7 @@
 
 namespace simdjson {
 namespace arm64 {
+namespace {
 
 // We sometimes call trailing_zero on inputs that are zero,
 // but the algorithms do not end up using the returned value.
@@ -64,6 +65,7 @@ really_inline bool mul_overflow(uint64_t value1, uint64_t value2, uint64_t *resu
 #endif
 }
 
+} // namespace {
 } // namespace arm64
 } // namespace simdjson