--- a/src/arm64/numberparsing.h
+++ b/src/arm64/numberparsing.h
@@ -1,9 +1,8 @@
 #ifndef SIMDJSON_ARM64_NUMBERPARSING_H
 #define SIMDJSON_ARM64_NUMBERPARSING_H
 
-namespace simdjson {
-namespace arm64 {
 namespace {
+namespace arm64 {
 
 // we don't have SSE, so let us use a scalar function
 // credit: https://johnnylee-sde.github.io/Fast-numeric-string-to-int/
@@ -15,9 +14,8 @@ static really_inline uint32_t parse_eight_digits_unrolled(const uint8_t *chars) 
   return uint32_t((val & 0x0000FFFF0000FFFF) * 42949672960001 >> 32);
 }
 
-} // namespace {
 } // namespace arm64
-} // namespace simdjson
+} // namespace {
 
 #define SWAR_NUMBER_PARSING