--- a/src/arm64/numberparsing.h
+++ b/src/arm64/numberparsing.h
@@ -3,6 +3,7 @@
 
 namespace simdjson {
 namespace arm64 {
+namespace {
 
 // we don't have SSE, so let us use a scalar function
 // credit: https://johnnylee-sde.github.io/Fast-numeric-string-to-int/
@@ -14,6 +15,7 @@ static really_inline uint32_t parse_eight_digits_unrolled(const uint8_t *chars) 
   return uint32_t((val & 0x0000FFFF0000FFFF) * 42949672960001 >> 32);
 }
 
+} // namespace {
 } // namespace arm64
 } // namespace simdjson