--- a/src/arm64/simd.h
+++ b/src/arm64/simd.h
@@ -9,6 +9,7 @@
 
 namespace simdjson {
 namespace arm64 {
+namespace {
 namespace simd {
 
 #ifdef SIMDJSON_REGULAR_VISUAL_STUDIO
@@ -494,6 +495,7 @@ really_inline int8x16_t make_int8x16_t(int8_t x1,  int8_t x2,  int8_t x3,  int8_
   }; // struct simd8x64<T>
 
 } // namespace simd
+} // namespace {
 } // namespace arm64
 } // namespace simdjson