--- a/src/generic/stage1/find_next_document_index.h
+++ b/src/generic/stage1/find_next_document_index.h
@@ -1,5 +1,6 @@
 namespace simdjson {
 namespace SIMDJSON_IMPLEMENTATION {
+namespace {
 
 /**
   * This algorithm is used to quickly identify the last structural position that
@@ -88,5 +89,6 @@ really_inline size_t trim_partial_utf8(const uint8_t *buf, size_t len) {
   return len;
 }
 
+} // namespace {
 } // namespace SIMDJSON_IMPLEMENTATION
 } // namespace simdjson