Adding one word to @jkeiser's PR boosts the GCC performance significantly in one test. The file twitter.json benefits a lot.

Performance effect on AMD Rome, GCC 9:

```
$ for i in ../jsonexamples/*.json; do echo $i; ./benchmark/parse -n 50 $i |grep GB|head -n 1; ../build_no_goto/benchmark/parse -n 50 $i |grep GB|head -n 1; done
../jsonexamples/apache_builds.json
|    Speed        :  29.3157 ns per block ( 98.65%) -   0.4581 ns per byte -   4.7164 ns per structural -   2.1828 GB/s
|    Speed        :  29.5123 ns per block ( 98.77%) -   0.4612 ns per byte -   4.7480 ns per structural -   2.1682 GB/s
../jsonexamples/canada.json
|    Speed        :  89.2029 ns per block ( 99.63%) -   1.3938 ns per byte -   9.3834 ns per structural -   0.7175 GB/s
|    Speed        :  88.8588 ns per block ( 99.63%) -   1.3884 ns per byte -   9.3472 ns per structural -   0.7202 GB/s
../jsonexamples/citm_catalog.json
|    Speed        :  26.4472 ns per block ( 99.37%) -   0.4132 ns per byte -   5.2486 ns per structural -   2.4199 GB/s
|    Speed        :  26.5883 ns per block ( 99.48%) -   0.4154 ns per byte -   5.2766 ns per structural -   2.4070 GB/s
../jsonexamples/github_events.json
|    Speed        :  27.9499 ns per block ( 98.25%) -   0.4369 ns per byte -   6.1124 ns per structural -   2.2891 GB/s
|    Speed        :  28.2750 ns per block ( 98.48%) -   0.4419 ns per byte -   6.1835 ns per structural -   2.2628 GB/s
../jsonexamples/google_maps_api_compact_response.json
|    Speed        :  66.8324 ns per block ( 97.01%) -   1.0467 ns per byte -   3.9679 ns per structural -   0.9554 GB/s
|    Speed        :  67.3730 ns per block ( 98.11%) -   1.0552 ns per byte -   4.0000 ns per structural -   0.9477 GB/s
../jsonexamples/google_maps_api_response.json
|    Speed        :  34.2059 ns per block ( 98.88%) -   0.5347 ns per byte -   4.4788 ns per structural -   1.8703 GB/s
|    Speed        :  34.4755 ns per block ( 96.91%) -   0.5389 ns per byte -   4.5141 ns per structural -   1.8557 GB/s
../jsonexamples/gsoc-2018.json
|    Speed        :  21.2044 ns per block ( 99.36%) -   0.3313 ns per byte -  14.5383 ns per structural -   3.0182 GB/s
|    Speed        :  21.0894 ns per block ( 99.49%) -   0.3295 ns per byte -  14.4595 ns per structural -   3.0347 GB/s
../jsonexamples/instruments.json
|    Speed        :  33.3677 ns per block ( 98.81%) -   0.5214 ns per byte -   4.2281 ns per structural -   1.9180 GB/s
|    Speed        :  34.4621 ns per block ( 98.97%) -   0.5385 ns per byte -   4.3667 ns per structural -   1.8571 GB/s
../jsonexamples/marine_ik.json
|    Speed        :  82.4708 ns per block ( 99.09%) -   1.2886 ns per byte -   5.9790 ns per structural -   0.7760 GB/s
|    Speed        :  83.6998 ns per block ( 99.65%) -   1.3078 ns per byte -   6.0681 ns per structural -   0.7646 GB/s
../jsonexamples/mesh.json
|    Speed        :  85.3834 ns per block ( 99.16%) -   1.3342 ns per byte -   6.2988 ns per structural -   0.7495 GB/s
|    Speed        :  86.5636 ns per block ( 98.71%) -   1.3527 ns per byte -   6.3858 ns per structural -   0.7393 GB/s
../jsonexamples/mesh.pretty.json
|    Speed        :  51.5500 ns per block ( 98.72%) -   0.8055 ns per byte -   8.2895 ns per structural -   1.2415 GB/s
|    Speed        :  51.7865 ns per block ( 99.07%) -   0.8092 ns per byte -   8.3275 ns per structural -   1.2358 GB/s
../jsonexamples/numbers.json
|    Speed        :  74.1633 ns per block ( 98.75%) -   1.1590 ns per byte -   8.6985 ns per structural -   0.8628 GB/s
|    Speed        :  73.7741 ns per block ( 99.19%) -   1.1529 ns per byte -   8.6528 ns per structural -   0.8674 GB/s
../jsonexamples/random.json
|    Speed        :  45.3791 ns per block ( 99.31%) -   0.7091 ns per byte -   4.1128 ns per structural -   1.4102 GB/s
|    Speed        :  48.3607 ns per block ( 99.14%) -   0.7557 ns per byte -   4.3830 ns per structural -   1.3233 GB/s
../jsonexamples/repeat.json
|    Speed        :  30.2191 ns per block ( 98.61%) -   0.4737 ns per byte -   5.2943 ns per structural -   2.1112 GB/s
|    Speed        :  30.5618 ns per block ( 98.35%) -   0.4790 ns per byte -   5.3543 ns per structural -   2.0875 GB/s
../jsonexamples/tree-pretty.json
|    Speed        :  33.8037 ns per block ( 96.52%) -   0.5282 ns per byte -   4.6119 ns per structural -   1.8931 GB/s
|    Speed        :  35.6037 ns per block ( 98.66%) -   0.5564 ns per byte -   4.8575 ns per structural -   1.7974 GB/s
../jsonexamples/twitter_api_compact_response.json
|    Speed        :  39.0633 ns per block ( 97.61%) -   0.6126 ns per byte -   5.0302 ns per structural -   1.6324 GB/s
|    Speed        :  40.5759 ns per block ( 98.58%) -   0.6363 ns per byte -   5.2249 ns per structural -   1.5715 GB/s
../jsonexamples/twitter_api_response.json
|    Speed        :  30.8912 ns per block ( 98.18%) -   0.4840 ns per byte -   5.1342 ns per structural -   2.0660 GB/s
|    Speed        :  32.2385 ns per block ( 95.83%) -   0.5051 ns per byte -   5.3581 ns per structural -   1.9796 GB/s
../jsonexamples/twitterescaped.json
|    Speed        :  58.9406 ns per block ( 99.07%) -   0.9210 ns per byte -   9.3730 ns per structural -   1.0858 GB/s
|    Speed        :  60.2780 ns per block ( 99.35%) -   0.9419 ns per byte -   9.5857 ns per structural -   1.0617 GB/s
../jsonexamples/twitter.json
|    Speed        :  30.3944 ns per block ( 97.36%) -   0.4749 ns per byte -   5.4275 ns per structural -   2.1055 GB/s
|    Speed        :  31.3792 ns per block ( 98.72%) -   0.4903 ns per byte -   5.6033 ns per structural -   2.0394 GB/s
../jsonexamples/twitter_timeline.json
|    Speed        :  46.7561 ns per block ( 97.86%) -   0.7307 ns per byte -   5.7940 ns per structural -   1.3686 GB/s
|    Speed        :  47.3303 ns per block ( 98.33%) -   0.7397 ns per byte -   5.8652 ns per structural -   1.3520 GB/s
../jsonexamples/update-center.json
|    Speed        :  37.7276 ns per block ( 99.05%) -   0.5895 ns per byte -   4.9561 ns per structural -   1.6963 GB/s
|    Speed        :  37.6495 ns per block ( 98.63%) -   0.5883 ns per byte -   4.9459 ns per structural -   1.6999 GB/s
```