@jkeiser 2020-08-01
Indeed! This one unlikely gives a massive benefit across the board vs. the original PR, pretty much sweeping away any performance degradation it had caused. skylake-x (vs. original PR, not master):

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 322.5 | 297.1 |    7% | 994.6 | 994.6 |    0% |  18522 |  18827 |   -1% |  20919 |  22265 |   -6% | 154706 | 154686 |    0% |
| gsoc-2018                        |  51997 |  74.8 |  72.6 |    2% | 171.2 | 171.5 |    0% |  31750 |  29440 |    7% |  53528 |  50858 |    4% | 168466 | 168378 |    0% |
| github_events                    |   1017 |  80.6 |  78.4 |    2% | 268.8 | 269.9 |    0% |    137 |    101 |   26% |      0 |      0 |    0% |      0 |      0 |    0% |
| random                           |   7976 | 144.2 | 140.6 |    2% | 495.9 | 497.3 |    0% |   2003 |   2042 |   -1% |    250 |      4 |   98% |  37828 |  37328 |    1% |
| instruments                      |   3442 | 104.4 | 102.0 |    2% | 376.1 | 375.6 |    0% |    516 |    492 |    4% |      4 |      1 |   75% |   2812 |    468 |   83% |
| twitter                          |   9867 |  94.7 |  92.8 |    2% | 298.6 | 299.6 |    0% |   3779 |   3453 |    8% |      0 |     21 |    0% |  31946 |  32157 |    0% |
| update-center                    |   8330 | 115.4 | 113.5 |    1% | 343.7 | 345.1 |    0% |   5575 |   5678 |   -1% |      9 |     22 | -144% |  33522 |  33809 |    0% |
| twitterescaped                   |   8787 | 193.1 | 190.6 |    1% | 513.6 | 511.1 |    0% |   5118 |   4915 |    3% |     58 |     14 |   75% |  30119 |  29130 |    3% |
| apache_builds                    |   1988 |  86.2 |  85.8 |    0% | 309.0 | 310.4 |    0% |    106 |    108 |   -1% |      0 |      0 |    0% |      0 |      7 |    0% |
| mesh.pretty                      |  24646 | 180.9 | 180.9 |    0% | 586.8 | 586.8 |    0% |  13481 |  13661 |   -1% |   1323 |   1276 |    3% |  88220 |  88150 |    0% |
| marine_ik                        |  46616 | 291.8 | 292.1 |    0% | 923.2 | 922.9 |    0% |  54351 |  53443 |    1% | 109139 | 108135 |    0% | 255953 | 255862 |    0% |
| numbers                          |   2345 | 257.4 | 258.7 |    0% | 802.3 | 802.3 |    0% |   2199 |   2176 |    1% |      0 |      4 |    0% |    126 |     27 |   78% |
| citm_catalog                     |  26987 |  84.5 |  87.3 |   -3% | 313.9 | 317.3 |   -1% |   1748 |   1783 |   -2% |    934 |    681 |   27% |  89500 |  89497 |    0% |
| mesh                             |  11306 | 294.7 | 305.9 |   -3% | 991.9 | 991.9 |    0% |  10030 |  10418 |   -3% |     14 |     55 | -292% |  60153 |  61000 |   -1% |