@jkeiser 2020-08-01
We get a more consistent, but less impressive, benefit if we apply it to the check to make sure the key is a string:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 322.5 | 306.7 |    4% | 994.6 | 991.5 |    0% |  18522 |  18880 |   -1% |  20919 |  22630 |   -8% | 154706 | 154736 |    0% |
| instruments                      |   3442 | 104.4 | 100.1 |    4% | 376.1 | 374.1 |    0% |    516 |    539 |   -4% |      4 |      2 |   50% |   2812 |    428 |   84% |
| gsoc-2018                        |  51997 |  74.8 |  71.8 |    4% | 171.2 | 171.2 |    0% |  31750 |  30446 |    4% |  53528 |  53026 |    0% | 168466 | 168456 |    0% |
| numbers                          |   2345 | 257.4 | 248.0 |    3% | 802.3 | 798.1 |    0% |   2199 |   2161 |    1% |      0 |      0 |    0% |    126 |     52 |   58% |
| github_events                    |   1017 |  80.6 |  77.9 |    3% | 268.8 | 269.0 |    0% |    137 |    122 |   10% |      0 |      0 |    0% |      0 |      0 |    0% |
| twitter                          |   9867 |  94.7 |  91.7 |    3% | 298.6 | 298.5 |    0% |   3779 |   3414 |    9% |      0 |      5 |    0% |  31946 |  33472 |   -4% |
| random                           |   7976 | 144.2 | 139.7 |    3% | 495.9 | 495.3 |    0% |   2003 |   2021 |    0% |    250 |      0 |    0% |  37828 |  36708 |    2% |
| twitterescaped                   |   8787 | 193.1 | 188.8 |    2% | 513.6 | 509.8 |    0% |   5118 |   4977 |    2% |     58 |      7 |   87% |  30119 |  29786 |    1% |
| update-center                    |   8330 | 115.4 | 113.0 |    2% | 343.7 | 343.6 |    0% |   5575 |   5787 |   -3% |      9 |      0 |    0% |  33522 |  33493 |    0% |
| mesh.pretty                      |  24646 | 180.9 | 177.5 |    1% | 586.8 | 585.5 |    0% |  13481 |  13170 |    2% |   1323 |    992 |   25% |  88220 |  88172 |    0% |
| marine_ik                        |  46616 | 291.8 | 288.4 |    1% | 923.2 | 920.0 |    0% |  54351 |  53981 |    0% | 109139 | 111907 |   -2% | 255953 | 255866 |    0% |
| apache_builds                    |   1988 |  86.2 |  85.7 |    0% | 309.0 | 309.5 |    0% |    106 |    113 |   -6% |      0 |      0 |    0% |      0 |     62 |    0% |
| mesh                             |  11306 | 294.7 | 297.0 |    0% | 991.9 | 989.0 |    0% |  10030 |  10108 |    0% |     14 |      5 |   64% |  60153 |  60919 |   -1% |
| citm_catalog                     |  26987 |  84.5 |  86.2 |   -2% | 313.9 | 316.7 |    0% |   1748 |   1732 |    0% |    934 |   1222 |  -30% |  89500 |  89393 |    0% |