@jkeiser 2020-08-01
And by replacing `if ( parser.start_object(false) ) { goto error; } else { goto object_begin; }` and friends with `if ( parser.start_object(false) ) { goto error; }; goto object_begin;` it gets even better:

+ ssh skylake-x cat /home/jkeiser/simdjson-remote/build/output.txt
| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 322.5 | 293.3 |    9% | 994.6 | 994.6 |    0% |  18522 |  18514 |    0% |  20919 |  21676 |   -3% | 154706 | 154764 |    0% |
| gsoc-2018                        |  51997 |  74.8 |  72.2 |    3% | 171.2 | 170.9 |    0% |  31750 |  30191 |    4% |  53528 |  50749 |    5% | 168466 | 168366 |    0% |
| instruments                      |   3442 | 104.4 | 100.9 |    3% | 376.1 | 374.6 |    0% |    516 |    484 |    6% |      4 |      3 |   25% |   2812 |   3169 |  -12% |
| marine_ik                        |  46616 | 291.8 | 283.0 |    3% | 923.2 | 922.4 |    0% |  54351 |  53948 |    0% | 109139 | 114611 |   -5% | 255953 | 255840 |    0% |
| github_events                    |   1017 |  80.6 |  78.3 |    2% | 268.8 | 267.8 |    0% |    137 |    116 |   15% |      0 |      0 |    0% |      0 |      0 |    0% |
| twitter                          |   9867 |  94.7 |  92.2 |    2% | 298.6 | 297.4 |    0% |   3779 |   3477 |    7% |      0 |    442 |    0% |  31946 |  32478 |   -1% |
| numbers                          |   2345 | 257.4 | 251.1 |    2% | 802.3 | 802.3 |    0% |   2199 |   2158 |    1% |      0 |      0 |    0% |    126 |      1 |   99% |
| random                           |   7976 | 144.2 | 141.2 |    2% | 495.9 | 493.9 |    0% |   2003 |   2007 |    0% |    250 |      7 |   97% |  37828 |  36818 |    2% |
| mesh.pretty                      |  24646 | 180.9 | 177.2 |    2% | 586.8 | 586.8 |    0% |  13481 |  13561 |    0% |   1323 |   1450 |   -9% |  88220 |  88110 |    0% |
| twitterescaped                   |   8787 | 193.1 | 190.3 |    1% | 513.6 | 512.3 |    0% |   5118 |   5054 |    1% |     58 |    120 | -106% |  30119 |  30651 |   -1% |
| apache_builds                    |   1988 |  86.2 |  85.1 |    1% | 309.0 | 308.1 |    0% |    106 |    107 |    0% |      0 |      0 |    0% |      0 |      0 |    0% |
| update-center                    |   8330 | 115.4 | 114.2 |    1% | 343.7 | 342.2 |    0% |   5575 |   5692 |   -2% |      9 |      1 |   88% |  33522 |  33601 |    0% |
| mesh                             |  11306 | 294.7 | 293.7 |    0% | 991.9 | 991.9 |    0% |  10030 |  10240 |   -2% |     14 |    498 | -3457% |  60153 |  59905 |    0% |
| citm_catalog                     |  26987 |  84.5 |  85.8 |   -1% | 313.9 | 313.4 |    0% |   1748 |   1743 |    0% |    934 |   1112 |  -19% |  89500 |  89402 |    0% |

Replacing all error conditions with unlikely makes things worse, FYI. I'll see what happens if we tag all places we first *encounter* errors with unlikely, but not tag `if (f()) { goto error; }` "rethrows."